move maps to component/map directory
