+++ src/app/components/maps/map-lockdown/map-lockdown.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, Input } from '@angular/core';
 import * as L from 'leaflet';
-import { MarkerService } from '../../_service/map/marker.service';
-import { ShapeService } from '../../_service/map/shape.service';
-import { MapTilesService } from '../../_service/map/map-tiles.service';
+import { MarkerService } from '../../../_service/map/marker.service';
+import { ShapeService } from '../../../_service/map/shape.service';
+import { MapTilesService } from '../../../_service/map/map-tiles.service';
 
 interface Country{
   name: string,
@@ -50,9 +50,9 @@ const colors = {
 })
 export class MapLockdownComponent implements OnInit {
 
-  public map: L.map;
-  public info: L.control;
-  public legend: L.control;
+  private map: L.map;
+  private info: L.control;
+  private legend: L.control;
   @Input() countries: Array<Country>;
 
   constructor(
